+++ src/model/person/boss/boss.h
@@ -19,7 +19,7 @@ public:
      * @param taxID the taxpayer identification number
      * @param credential the login credentials
      */
-    explicit Boss(std::string name = "Boss", int taxID = Person::DEFAULT_TAX_ID,
+    explicit Boss(std::string name = "Boss", unsigned long taxID = Person::DEFAULT_TAX_ID,
                   Credential credential = {DEFAULT_USERNAME, DEFAULT_PASSWORD});
 
     /**
